fix multiple documento controller
